+++ fet/src/engine/rules.h
@@ -479,7 +479,7 @@ public:
 		//int _parities[],
 		bool _active[],
 		int _minDayDistance,
-		int _weightPercentage,
+		double _weightPercentage,
 		bool _consecutiveIfSameDay,
 		//int _preferredDays[],
 		//int _preferredHours[],
